irqchip: irq-renesas-rzt2h: move negative request number handling
Move it into rzt2h_icu_register_dma_req(), turning it into an alias for
rzt2h_icu_register_dma_req_new().

+++ include/linux/irqchip/irq-renesas-rzt2h.h
@@ -14,7 +14,7 @@
 
 #ifdef CONFIG_RENESAS_RZT2H_ICU
 void rzt2h_icu_register_dma_req(struct platform_device *icu_dev, u8 dmac_index, u8 dmac_channel,
-				u16 req_no);
+				int req_no);
 void rzt2h_icu_register_dma_req_new(struct platform_device *icu_dev, u8 dmac_index, u8 dmac_channel,
 				    int req_no);
 #else
